AgBiz-Logic-dev/requirements.txt
coverage==4.0.3
django==2.2.7
django-admin-bootstrapped==2.5.6
django-bootstrap-form==3.2
django-crispy-forms==1.8.0
django-extra-views==0.7.1
django-filter==0.13.0
django-formtools==2.1
django-localflavor==2.2
django-nose==1.4.6
django-widget-tweaks==1.4.1
djangorestframework==3.10.3
Markdown==2.6.6
nose==1.3.7
selenium==2.52.0
setuptools==18.2
six==1.10.0
wheel==0.24.0
xlrd==0.9.4
mock==2.0.0
openpyxl==2.4.0
paho-mqtt
psycopg2-binary==2.8.6
numpy
geopy
netcdf4
requests==2.18.4
